Weight allowed to be a float.

+++ recengine/cli.py
@@ -16,7 +16,7 @@ def hello():
 
 @cli.command()
 @click.option('--age', '-a', prompt=True, required=True, type=int)
-@click.option('--weight', '-w', prompt=True, required=True, type=int)
+@click.option('--weight', '-w', prompt=True, required=True, type=float)
 @click.option('--performance', '-p', prompt=True, required=True, type=float)
 @click.option('--sex', type=click.Choice(['MAN', 'WOMAN', 'OTHER'], case_sensitive=False),
               prompt=True, required=True)
